+++ JudoPayDotNet/Models/GetWebPaymentResponseModel.cs
@@ -83,7 +83,10 @@ namespace JudoPayDotNet.Models
         [DataMember(EmitDefaultValue = false)]
         public string ShortReference { get; private set; }
 
+        /// <summary>
+        /// Set true for consumer initiated preAuths that can be incremented before they are fully collected
+        /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public bool? DelayedAuthorisation { get; private set; }
+        public bool? AllowIncrement { get; set; }
     }
 }