+++ JudoPayDotNet/Models/WebPaymentRequestModel.cs
@@ -141,11 +141,10 @@ namespace JudoPayDotNet.Models
         public PrimaryAccountDetailsModel PrimaryAccountDetails { get; set; }
 
         /// <summary>
-        /// Only set this flag on payment sessions for PreAuths, to authenticate a card holder with 3D secure without
-        /// performing payment authorisation.
+        /// Set true for consumer initiated preAuths that can be incremented before they are fully collected
         /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public bool? DelayedAuthorisation { get; set; }
+        [DataMember(IsRequired = false, EmitDefaultValue = false)]
+        public bool? AllowIncrement { get; set; }
 
         private Dictionary<string, string> _httpHeaders;